Added components folder for move to Redux
